requirements.txt
# requirements.txt ファイル

streamlit==1.25.0
streamlit-authenticator==0.2.1
google-cloud-texttospeech==2.15.0
google-api-python-client==2.97.0
openai==0.27.10
requests==2.31.0
numpy==1.26.0
pandas==2.1.2
plotly==5.17.0
seaborn==0.13.0
matplotlib==3.9.0
pydub==0.25.1
Pillow==10.0.1
exifread==3.2.1
pdfplumber==0.9.0
python-docx==0.8.11
bcrypt==4.0.1
pyyaml==6.0
streamlit-lottie==0.0.8
lxml==4.9.3
scipy==1.11.3
SQLAlchemy==2.1.1
psycopg2-binary==2.9.6
pyjwt==2.8.0
tensorflow==2.14.0
torch==2.1.0
transformers==4.36.1
